+++ core/ExceptionHandler.js
@@ -4,8 +4,11 @@
  *
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
-import {Exception, stringifyErrorSafely} from '@xh/hoist/exception';
+import {Exception} from '@xh/hoist/exception';
+import {fragment, span} from '@xh/hoist/cmp/layout';
 import {stripTags} from '@xh/hoist/utils/js';
+import {Icon} from '@xh/hoist/icon';
+import {forOwn, has, isArray, isNil, isObject, omitBy, set} from 'lodash';
 import {XH} from './XH';
 
 /**
@@ -13,6 +16,29 @@ import {XH} from './XH';
  * Manages the logging and display of exceptions.
  */
 export class ExceptionHandler {
+
+    /**
+     * Property paths within error details JSON to replace with '******'
+     * @type {string[]}
+     */
+    static REDACT_PATHS = [
+        'fetchOptions.headers.Authorization'
+    ];
+
+    /**
+     * Default type of alert to use to display exceptions with `showAlert`.
+     * Valid options are 'dialog'|'toast'.
+     * @type {string}
+     */
+    static ALERT_TYPE = 'dialog';
+
+    /**
+     * Default props provided to toast, when alert type is 'toast'
+     * @type {Object}
+     */
+    static TOAST_PROPS = {
+        timeout: 10000
+    };
 
     #isUnloading = false;
 
@@ -50,6 +76,8 @@ export class ExceptionHandler {
      *      'isAutoRefresh' fetch exceptions.
      * @param {boolean} [options.showAlert] - display an alert dialog to the user. Default true,
      *      excepting 'isAutoRefresh' and 'isFetchAborted' exceptions.
+     * @param {string} [options.alertType] - if `showAlert`, which type of alert to display.
+     *      Valid options are 'dialog'|'toast'. Defaults to ExceptionHandler.ALERT_TYPE.
      * @param {boolean} [options.requireReload] - force user to fully refresh the app in order to
      *      dismiss - default false, excepting session-related exceptions.
      * @param {string[]} [options.hideParams] - A list of parameters that should be hidden from
@@ -60,12 +88,29 @@ export class ExceptionHandler {
 
         ({exception, options} = this.parseArgs(exception, options));
 
+        const {showAlert, alertType, logOnServer, title, message, showAsError} = options;
+
         this.logException(exception, options);
-        if (options.showAlert) {
-            XH.appContainerModel.exceptionDialogModel.show(exception, options);
+        if (showAlert) {
+            if (alertType === 'toast') {
+                XH.toast({
+                    message: fragment(
+                        span({className: 'xh-toast__title', item: title, omit: !title}),
+                        span({className: 'xh-toast__body', item: message})
+                    ),
+                    actionButtonProps: {
+                        icon: Icon.search(),
+                        onClick: () => XH.appContainerModel.exceptionDialogModel.show(exception, options)
+                    },
+                    intent: showAsError ? 'danger' : 'primary',
+                    ...ExceptionHandler.TOAST_PROPS
+                });
+            } else {
+                XH.appContainerModel.exceptionDialogModel.show(exception, options);
+            }
         }
-        if (options.logOnServer) {
-            this.logOnServerAsync({exception, userAlerted: options.showAlert});
+        if (logOnServer) {
+            this.logOnServerAsync({exception, userAlerted: showAlert});
         }
     }
 
@@ -110,7 +155,7 @@ export class ExceptionHandler {
      */
     async logOnServerAsync({exception, userAlerted, userMessage}) {
         try {
-            const error = stringifyErrorSafely(exception),
+            const error = this.stringifyErrorSafely(exception),
                 username = XH.getUsername();
 
             if (!username) {
@@ -136,6 +181,60 @@ export class ExceptionHandler {
         }
     }
 
+    /**
+     * Serialize an error object safely for submission to server, or user display.
+     * This method will avoid circular references and will trim the depth of the object.
+     *
+     * @param {Error} error
+     * @return string
+     */
+    stringifyErrorSafely(error) {
+        try {
+            // 1) Create basic structure.
+            // Raw Error does not have 'own' properties, so be explicit about core name/message/stack
+            // Order here intentional for serialization
+            let ret = {
+                name: error.name,
+                message: error.message
+            };
+            Object.assign(ret, error);
+            ret.stack = error.stack?.split(/\n/g);
+
+            ret = omitBy(ret, isNil);
+
+            // 2) Deep clone/protect against circularity/monstrosity
+            ret = this.cloneAndTrim(ret);
+
+            // 3) Additional ad-hoc cleanups
+            // Remove noisy grails exception wrapper info
+            // Remove verbose loadSpec from fetchOptions
+            const {serverDetails} = ret;
+            if (serverDetails?.className === 'GrailsCompressingFilter') {
+                delete serverDetails.className;
+                delete serverDetails.lineNumber;
+            }
+
+            const {fetchOptions} = ret;
+            if (fetchOptions?.loadSpec) {
+                fetchOptions.loadType = fetchOptions.loadSpec.typeDisplay;
+                fetchOptions.loadNumber = fetchOptions.loadSpec.loadNumber;
+                delete fetchOptions.loadSpec;
+            }
+
+            // 4) Redact specified values
+            ExceptionHandler.REDACT_PATHS.forEach(path => {
+                if (has(ret, path)) set(ret, path, '******');
+            });
+
+            // 5) Stringify and cleanse
+            return stripTags(JSON.stringify(ret, null, 4));
+        } catch (e) {
+            const message = 'Failed to serialize error';
+            console.error(message, error, e);
+            return JSON.stringify({message}, null, 4);
+        }
+    }
+
     //--------------------------------
     // Implementation
     //--------------------------------
@@ -154,7 +253,6 @@ export class ExceptionHandler {
 
         return {exception, options};
     }
-
 
     hideParams(exception, options) {
         const {fetchOptions} = exception,
@@ -186,6 +284,7 @@ export class ExceptionHandler {
         ret.logOnServer = ret.logOnServer ?? (ret.showAsError && !isAutoRefresh);
         ret.showAlert = ret.showAlert ?? (!isAutoRefresh && !isFetchAborted);
         ret.requireReload = ret.requireReload ?? !!e.requireReload;
+        ret.alertType = ret.alertType ?? ExceptionHandler.ALERT_TYPE;
 
         ret.title = ret.title || (ret.showAsError ? 'Error' : 'Alert');
         ret.message = ret.message || e.message || e.name || 'An unknown error occurred.';
@@ -223,5 +322,28 @@ export class ExceptionHandler {
         // statuses of 0, 4XX, 5XX are server errors, so the javascript stack
         // is irrelevant and potentially misleading
         if (/^[045]/.test(exception.httpStatus)) delete exception.stack;
+    }
+
+    cloneAndTrim(obj, depth = 5) {
+        // Create a depth-constrained, deep copy of an object for safe-use in stringify
+        //   - Skip private _XXXX properties.
+        //   - Don't touch objects that implement toJSON()
+        if (depth < 1) return null;
+
+        const ret = {};
+        forOwn(obj, (val, key) => {
+            if (key.startsWith('_')) return;
+            if (val && !val.toJSON) {
+                if (isObject(val)) {
+                    val = depth > 1 ? this.cloneAndTrim(val, depth - 1) : '{...}';
+                }
+                if (isArray(val)) {
+                    val = depth > 1 ? val.map(it => this.cloneAndTrim(it, depth - 1)) : '[...]';
+                }
+            }
+            ret[key] = val;
+        });
+
+        return ret;
     }
 }