+++ CHANGELOG.md
@@ -3,11 +3,37 @@
 ## v46.0.0-SNAPSHOT - unreleased
 
 ### 🎁 New Features
+
+* `ExceptionHandler` provides a collection of overwritable static properties, allowing
+  you to set app-wide default behaviour for exception handling.
+* `XH.handleException()` has new `alertType` option, enabling you to show as either `dialog` or `toast`.
+* `XH.toast()` takes new `actionButtonProps` option, for showing an action button within the toast.
+
+### 💥 Breaking Changes
+
+* `LocalDate` methods `toString()`, `toJSON()`, `valueOf()`, and `isoString()` now all return the
+  standard ISO format: `YYYY-MM-DD`. (Prior versions returned`YYYYMMDD`).
+  This is consistent with the built-in javascript `Date.toISOString()`
+* The `stringifyErrorSafely` function has been moved from the `@xh/hoist/exception` package to a
+  public method on `XH.exceptionHandler`.
+    * Not expected to have much or any impact on application code.
+
+### ⚙️ Technical
+* New flag `GridModel.isInEditingMode`
+
+## v45.0.2 - 2022-01-13
+
+### 🎁 New Features
+
 * `FilterChooser` has new `menuWidth` prop, allowing you to specify as width for the dropdown
   menu that is different from the control.
 
 ### 🐞 Bug Fixes
+
 * Fixed cache clearing method on Admin Console's Server > Services tab.
+* Several fixes to behavior of `GridAutosizeMode.MANAGED`
+
+[Commit Log](https://github.com/xh/hoist-react/compare/v45.0.1...v45.0.2)
 
 ## v45.0.1 - 2022-01-07